refactor: Refactor CompanyDataStorage handling
Streamline CompanyDataStorage operations by extracting functionality into a new CompanyDataStorageHandler class. This improves code readability and maintainability by consolidating repetitive storage interactions and updating methods to better organize and handle company data.
