+++ apps/web/src/utils/track/TrackSiret.ts
@@ -2,7 +2,6 @@ import EstablishmentApi from '@/service/api/establishmentApi'
 import type { Track, TrackOptionItem, TrackOptionsUnion } from '@/types'
 import { EstablishmentFront, SiretValue } from '@/types'
 import { SiretValidator } from '@tee/common'
-import CompanyDataStorage from '@/utils/storage/companyDataStorage'
 
 export default class TrackSiret {
   static async search(query: string) {
@@ -16,10 +15,6 @@ export default class TrackSiret {
     remove = false,
     forceKeep = false
   ): TrackOptionItem {
-    if (questionnaireData) {
-      this.saveData(option, questionnaireData)
-    }
-
     return {
       option: {
         ...option,
@@ -29,12 +24,6 @@ export default class TrackSiret {
       remove: remove,
       forceKeep: forceKeep
     }
-  }
-
-  static saveData(option: TrackOptionsUnion, questionnaireData?: EstablishmentFront) {
-    const data = questionnaireData || option.questionnaireData
-
-    CompanyDataStorage.setSiret(data as EstablishmentFront)
   }
 
   static async getOptionBySiret(track: Track, siret: string): Promise<TrackOptionsUnion | undefined> {
